+++ src/widget/downlist/popup.downlist.js
@@ -118,7 +118,7 @@ BI.DownListPopup = BI.inherit(BI.Pane, {
                         var fatherValue = BI.deepClone(item.el.value);
                         var childValue = BI.deepClone(child.value);
                         self.singleValues.push(child.value);
-                        child.type = "bi.down_list_item";
+                        child.type = child.type || "bi.down_list_item";
                         child.extraCls = " child-down-list-item";
                         child.title = child.title || child.text;
                         child.textRgap = 10;