+++ changelog.md
@@ -1,5 +1,6 @@
 # 更新日志
 2.0(2020-07)
+- 修复了datePicker在setValue的时候没有动态刷新可用月份的问题
 - 同步复选下拉及其面板新增getAllValue获取所有已选值
 - 同步复选下拉树及其面板新增getAllValue获取完整的选中树节点
 - 修复date_picker最大值最小值与面板展示判断问题